+++ test/benchmark/engine_preparation.py
@@ -9,7 +9,7 @@ from fastapi.testclient import TestClient
 
 from voicevox_engine.app.application import generate_app
 from voicevox_engine.core.core_initializer import initialize_cores
-from voicevox_engine.engine_manifest.engine_manifest import load_manifest
+from voicevox_engine.engine_manifest.engine_manifest_loader import load_manifest
 from voicevox_engine.preset.PresetManager import PresetManager
 from voicevox_engine.setting.Setting import SettingHandler
 from voicevox_engine.tts_pipeline.tts_engine import make_tts_engines_from_cores
